feat: separate photo related calls from createSubmission
